Finishing documentation; cleanup.

+++ Sensors/SMCBatteryManager/BatteryManagerState.hpp
@@ -55,7 +55,7 @@ struct BatteryInfo {
 		uint32_t designCapacityWarning {0};
 		uint32_t designCapacityLow {0};
 		uint16_t temperatureRaw {0};
-		uint16_t temperature {0};
+		double temperature {0};
 		bool powerUnitIsWatt {false};
 		bool calculatedACAdapterConnected {false};
 		bool bad {false};
